docs(stories): consolidate banner harness controls and improve storybook config
- Add showSeverityIcon, showCloseButton, duration, iconSize, iconAnimation
  controls to every overlay/actions/custom story so the Controls panel
  covers TbxMatBannerConfig's full surface.
- Extract the overlay and actions-group harnesses (plus shared types,
  argTypes, defaults, and the SVG-icon provider decorator) into
  notification-style *.stories.common modules so dev and docs stories
  share one source of truth.
- Collapse banner-icon-variants from six baked-in stories (Default,
  Large, StateTransition, LargeStateTransition, Pulse, LargePulse)
  into one story with iconSize + iconAnimation knobs.
- Drop the redundant "Position: Bottom" trigger row in dev/Overlay; the
  verticalPosition control already covers it.
- Add iconSize / iconAnimation controls to dev/Inline (previously had
  zero argTypes).
- Preload Material Symbols Rounded with display=block plus a fonts-loaded
  visibility gate so severity and close icons render glyphs from first
  paint instead of flashing the ligature text.
- Disable save-from-controls UI, reset story args on each navigation, and
  hide the What's New notifications panel.
