+++ .storybook/preview-head.html
@@ -1,4 +1,25 @@
 <link rel="preconnect" href="https://fonts.googleapis.com" />
 <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+<link rel="preload" as="style" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Rounded:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200&display=block" />
 <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;600;700&display=swap" rel="stylesheet" />
-<link href="https://fonts.googleapis.com/css2?family=Material+Symbols+Rounded:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" rel="stylesheet" />
+<link href="https://fonts.googleapis.com/css2?family=Material+Symbols+Rounded:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200&display=block" rel="stylesheet" />
+<style>
+    /* Hide Material Symbols ligature text until the font is loaded so the
+       severity/close icons don't flash as raw text on first paint. */
+    .material-symbols-rounded {
+        font-display: block;
+        visibility: hidden;
+    }
+    .fonts-loaded .material-symbols-rounded {
+        visibility: visible;
+    }
+</style>
+<script>
+    if (document.fonts && document.fonts.load) {
+        document.fonts.load('24px "Material Symbols Rounded"').then(function () {
+            document.documentElement.classList.add("fonts-loaded");
+        });
+    } else {
+        document.documentElement.classList.add("fonts-loaded");
+    }
+</script>
