+++ .storybook/preview.ts
@@ -53,6 +53,7 @@ const preview: Preview = {
             },
         },
         controls: {
+            disableSaveFromUI: true,
             matchers: {
                 color: /(background|color)$/i,
                 date: /Date$/i,
