+++ src/components/banner-icon-variants.stories.ts
@@ -1,181 +1,25 @@
-import { Component, inject, input } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
 import type { Meta, StoryObj } from '@storybook/angular';
 import { moduleMetadata } from '@storybook/angular';
-import { TbxMatBannerAnimation } from '../enums/banner-animation.enum';
-import { TbxMatBannerService } from '../services/banner.service';
-import { withCustomProperties, withDefaultProperties } from './story-overrides';
+import { BannerOverlayHarnessComponent, DEFAULT_OVERLAY_ARGS, SHARED_OVERLAY_ARG_TYPES } from './banner-overlay.stories.common';
 
-// ─── CSS Custom Property Overrides ───────────────────────────────────────────
-
-const LARGE_ICON_CSS = `
-    html {
-        --tbx-mat-banner-icon-size: 3rem;
-    }
-`;
-
-const STATE_TRANSITION_CSS = `
-    @keyframes tbx-banner-icon-fill {
-        from { font-variation-settings: 'FILL' 0, 'wght' 400, 'GRAD' 0, 'opsz' 24; }
-        to   { font-variation-settings: 'FILL' 1, 'wght' 400, 'GRAD' 0, 'opsz' 24; }
-    }
-    .tbx-mat-banner-overlay-panel .material-symbols-rounded {
-        animation: tbx-banner-icon-fill 0.3s ease-in-out 0.15s forwards;
-        font-variation-settings: 'FILL' 0, 'wght' 400, 'GRAD' 0, 'opsz' 24;
-    }
-`;
-
-const LARGE_ICON_STATE_TRANSITION_CSS = `
-    html {
-        --tbx-mat-banner-icon-size: 3rem;
-    }
-    @keyframes tbx-banner-icon-fill {
-        from { font-variation-settings: 'FILL' 0, 'wght' 400, 'GRAD' 0, 'opsz' 24; }
-        to   { font-variation-settings: 'FILL' 1, 'wght' 400, 'GRAD' 0, 'opsz' 24; }
-    }
-    .tbx-mat-banner-overlay-panel .material-symbols-rounded {
-        animation: tbx-banner-icon-fill 0.3s ease-in-out 0.15s forwards;
-        font-variation-settings: 'FILL' 0, 'wght' 400, 'GRAD' 0, 'opsz' 24;
-    }
-`;
-
-const PULSE_CSS = `
-    @keyframes tbx-banner-icon-pulse {
-        from { font-variation-settings: 'FILL' 0, 'wght' 400, 'GRAD' 0, 'opsz' 24; }
-        to   { font-variation-settings: 'FILL' 1, 'wght' 400, 'GRAD' 0, 'opsz' 24; }
-    }
-    .tbx-mat-banner-overlay-panel .tbx-mat-banner-icon {
-        animation: tbx-banner-icon-pulse 1s ease-in-out infinite alternate;
-        font-variation-settings: 'FILL' 0, 'wght' 400, 'GRAD' 0, 'opsz' 24;
-    }
-`;
-
-const LARGE_ICON_PULSE_CSS = `
-    html {
-        --tbx-mat-banner-icon-size: 3rem;
-    }
-    @keyframes tbx-banner-icon-pulse {
-        from { font-variation-settings: 'FILL' 0, 'wght' 400, 'GRAD' 0, 'opsz' 24; }
-        to   { font-variation-settings: 'FILL' 1, 'wght' 400, 'GRAD' 0, 'opsz' 24; }
-    }
-    .tbx-mat-banner-overlay-panel .tbx-mat-banner-icon {
-        animation: tbx-banner-icon-pulse 1s ease-in-out infinite alternate;
-        font-variation-settings: 'FILL' 0, 'wght' 400, 'GRAD' 0, 'opsz' 24;
-    }
-`;
-
-// ─── Harness Component ───────────────────────────────────────────────────────
-
-@Component({
-    selector: 'tbx-banner-icon-variants-harness',
-    imports: [MatButtonModule],
-    template: `
-        <div class="harness">
-            <h3>Severity Triggers</h3>
-            <div class="button-group">
-                <button mat-flat-button (click)="fire('default')">Default</button>
-                <button mat-flat-button (click)="fire('success')">Success</button>
-                <button mat-flat-button (click)="fire('error')">Error</button>
-                <button mat-flat-button (click)="fire('warning')">Warning</button>
-                <button mat-flat-button (click)="fire('information')">Information</button>
-                <button mat-flat-button (click)="fire('help')">Help</button>
-            </div>
-
-            <h3>Queue All</h3>
-            <div class="button-group">
-                <button mat-flat-button (click)="queueAll()">Fire 6 Queued</button>
-                <button mat-flat-button (click)="banner.dismissAll()">Dismiss All</button>
-            </div>
-            <p class="state">Active: {{ banner.isActive() }} &middot; Pending: {{ banner.pendingCount() }}</p>
-        </div>
-    `,
-    styleUrl: './story-harness.css',
-})
-class BannerIconVariantsHarnessComponent {
-    readonly banner = inject(TbxMatBannerService);
-    readonly verticalPosition = input<'top' | 'bottom'>('top');
-    readonly animation = input<TbxMatBannerAnimation>(TbxMatBannerAnimation.None);
-
-    private readonly messages: Record<string, string> = {
-        default: 'This is a default banner.',
-        success: 'Operation completed successfully.',
-        error: 'Something went wrong. Please try again.',
-        warning: 'Your session will expire in 5 minutes.',
-        information: 'A new version is available.',
-        help: 'Click the + button to add a new item.',
-    };
-
-    fire(level: string): void {
-        const method = this.banner[level as keyof TbxMatBannerService] as (msg: string, args?: object) => void;
-        method.call(this.banner, this.messages[level], {
-            verticalPosition: this.verticalPosition(),
-            animation: this.animation(),
-        });
-    }
-
-    queueAll(): void {
-        const args = { verticalPosition: this.verticalPosition(), animation: this.animation() };
-        this.banner.default('Step 1: This is a default banner.', args);
-        this.banner.success('Step 2: Operation completed successfully.', args);
-        this.banner.error('Step 3: Something went wrong.', args);
-        this.banner.warning('Step 4: Review needed.', args);
-        this.banner.information('Step 5: A new version is available.', args);
-        this.banner.help('Step 6: Click the + button to add a new item.', args);
-    }
-}
-
-// ─── Meta ────────────────────────────────────────────────────────────────────
-
-const meta: Meta<BannerIconVariantsHarnessComponent> = {
+const meta: Meta<BannerOverlayHarnessComponent> = {
     title: 'Banners/Overlay Font Icon Variants',
-    component: BannerIconVariantsHarnessComponent,
-    decorators: [moduleMetadata({ imports: [BannerIconVariantsHarnessComponent] })],
-    argTypes: {
-        verticalPosition: {
-            control: 'select',
-            options: ['top', 'bottom'],
-            description: 'Vertical position of the overlay banner',
-        },
-        animation: {
-            control: 'select',
-            options: [TbxMatBannerAnimation.None, TbxMatBannerAnimation.Slide, TbxMatBannerAnimation.Fade],
-            description: 'Enter/exit animation mode',
-        },
-    },
+    component: BannerOverlayHarnessComponent,
+    decorators: [moduleMetadata({ imports: [BannerOverlayHarnessComponent] })],
+    argTypes: SHARED_OVERLAY_ARG_TYPES,
 };
 
 export default meta;
-type Story = StoryObj<BannerIconVariantsHarnessComponent>;
+type Story = StoryObj<BannerOverlayHarnessComponent>;
 
-export const DefaultIcons: Story = {
-    name: 'Default Icons',
-    args: { verticalPosition: 'top', animation: TbxMatBannerAnimation.None },
-    decorators: [withDefaultProperties()],
-};
-
-export const LargeIcons: Story = {
-    name: 'Large Icons',
-    args: { verticalPosition: 'top', animation: TbxMatBannerAnimation.None },
-    decorators: [withCustomProperties(LARGE_ICON_CSS)],
-};
-
-export const StateTransition: Story = {
-    name: 'State Transition',
-    args: { verticalPosition: 'top', animation: TbxMatBannerAnimation.None },
-    decorators: [withCustomProperties(STATE_TRANSITION_CSS)],
-};
-
-export const LargeStateTransition: Story = {
-    name: 'Large + State Transition',
-    decorators: [withCustomProperties(LARGE_ICON_STATE_TRANSITION_CSS)],
-};
-
-export const Pulse: Story = {
-    name: 'Pulse',
-    decorators: [withCustomProperties(PULSE_CSS)],
-};
-
-export const LargePulse: Story = {
-    name: 'Large + Pulse',
-    decorators: [withCustomProperties(LARGE_ICON_PULSE_CSS)],
+export const Default: Story = {
+    name: 'Font Icon Variants',
+    args: { ...DEFAULT_OVERLAY_ARGS },
+    parameters: {
+        docs: {
+            description: {
+                story: 'Demonstrates the Material Symbols font icon size and animation knobs. Use the **Icon Size** and **Icon Animation** controls to compare standard / medium / large with `none`, `state-transition` (fill on enter), and `pulse`. Effects target the snackbar overlay container and the `.tbx-mat-banner-icon` element.',
+            },
+        },
+    },
 };
